feat: cta bottomSheet migration on AppInner Compo
